Refactor FieldMetadataEnumValidationService to centralize default value validation logic
